Break up config source files
